resume_builder_django/requirements.txt
asgiref==3.9.1
crispy-bootstrap4==2025.6
dj-database-url==3.0.1
Django==5.2.4
django-allauth @ git+https://github.com/pennersr/django-allauth.git@1159fd47b6a145d5dc2447ab44ee5bd45aa3a529
django-crispy-forms==2.4
psycopg2-binary==2.9.10
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
asgiref==3.9.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap4==2025.6
cryptography==45.0.5
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.0.14
django-allauth==65.10.0
django-crispy-forms==2.4
django-formtools==2.5.1
django-heroku==0.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
fonttools==4.58.5
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
lxml==6.0.0
mssql-django==1.5
oauthlib==3.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.3
pyphen==0.17.2
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.9.0
zopfli==0.2.3.post1
